# Enrich tailsamplingprocessor string-attribute-filter to support regex on span attribute filtering