**Description:** <Describe what has changed.>
Allow `string-attribute-filter` in `tailsamplingprocessor`  to support both plain text and `regex` rule when filtering the span attribute by matching the value. 
This is a task in P2 milestone list.

New Config Example,
```
processors:
  tail_sampling:
    decision_wait: 10s
    num_traces: 100
    expected_new_traces_per_sec: 10
    policies:
       [
          {
               name: test-policy-3,
               type: string_attribute,
               string_attribute: {key: "key", values: ["v[0-9]+\.Health\.Check$", "plain_text"]}
          }
      ]
```
### Benchmark
**Before,**
```
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/processor/tailsamplingprocessor/sampling
cpu: Intel(R) Core(TM) i7-8569U CPU @ 2.80GHz
**BenchmarkStringTagFilterEvaluatePlainText**
BenchmarkStringTagFilterEvaluatePlainText-8   	29896698	        40.31 ns/op
```
**After,**
```
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/processor/tailsamplingprocessor/sampling
cpu: Intel(R) Core(TM) i7-8569U CPU @ 2.80GHz
BenchmarkStringTagFilterEvaluatePlainText
BenchmarkStringTagFilterEvaluatePlainText-8   	30917172	        41.27 ns/op

pkg: github.com/open-telemetry/opentelemetry-collector-contrib/processor/tailsamplingprocessor/sampling
cpu: Intel(R) Core(TM) i7-8569U CPU @ 2.80GHz
BenchmarkStringTagFilterEvaluateRegex
BenchmarkStringTagFilterEvaluateRegex-8   	18788356	        63.94 ns/op
```

**Link to tracking Issue:** 
https://github.com/open-telemetry/opentelemetry-collector/issues/226

**Testing:**
`make all`
**Documentation:** 
README