@mxiamxia 2021-05-16
> LGTM, just had a comment about testability. If you decide to keep the code as-is, that's also fine by me and I'll change the review to "approved".

Thanks for all the feedbacks. I updated the godoc for `Evaluate()` func per the suggestion and leave the rest as-is.